Fix imports for moved threshold defs
