+++ src/parser/monk/brewmaster/modules/spells/PurifyingBrew.tsx
@@ -5,8 +5,7 @@ import SpellIcon from 'common/SpellIcon';
 import SpellLink from 'common/SpellLink';
 import SPELLS from 'common/SPELLS';
 import Analyzer, { SELECTED_PLAYER } from 'parser/core/Analyzer';
-import { When } from 'parser/core/ParseResults';
-import { ThresholdStyle } from 'parser/core/Thresholds';
+import { When, ThresholdStyle } from 'parser/core/ParseResults';
 import Events, { EventType, RemoveDebuffEvent, CastEvent } from 'parser/core/Events';
 import EventFilter from 'parser/core/EventFilter';
 import Abilities from 'parser/core/modules/Abilities';
