+++ src/parser/shared/modules/AlwaysBeCasting.tsx
@@ -4,8 +4,7 @@ import Icon from 'common/Icon';
 import { formatPercentage } from 'common/format';
 import Analyzer from 'parser/core/Analyzer';
 import { EndChannelEvent, EventType, GlobalCooldownEvent } from 'parser/core/Events';
-import { ThresholdStyle } from 'parser/core/Thresholds';
-import { When } from 'parser/core/ParseResults';
+import { When, ThresholdStyle } from 'parser/core/ParseResults';
 import StatisticBox, { STATISTIC_ORDER } from 'interface/others/StatisticBox';
 import Tooltip from 'common/Tooltip';
 import Abilities from '../../core/modules/Abilities';
